backend/requirements.txt
aniso8601==8.0.0
astroid==2.4.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.0
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
